+++ arithmetization/src/main/java/net/consensys/linea/zktracer/module/hub/state/Block.java
@@ -13,7 +13,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package net.consensys.linea.zktracer.module.hub.transients;
+package net.consensys.linea.zktracer.module.hub.state;
 
 import java.util.*;
